refactor: move tests to their own files
